requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes --format requirements-txt
# -e .
aniso8601==10.0.0
apispec==6.6.1
apispec-webframeworks==1.1.0
arrow==1.3.0
attrs==25.3.0
automat==25.4.16
azure-core==1.33.0
azure-identity==1.21.0
azure-storage-blob==12.25.1
azure-storage-file-datalake==12.20.0
bidict==0.23.1
binaryornot==0.4.4
blinker==1.9.0
boto3==1.34.113
botocore==1.34.162
certifi==2025.1.31
cffi==1.17.1 ; platform_python_implementation != 'PyPy'
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6 ; platform_system == 'Windows'
constantly==23.10.4
cookiecutter==2.6.0
cryptography==44.0.2
deepdiff==7.0.1
dnspython==2.7.0
et-xmlfile==2.0.0
flask==3.1.0
flask-cors==5.0.1
flask-restful==0.3.10
flask-socketio==5.4.1
gevent==24.11.1
gevent-websocket==0.10.1
gitignore-parser==0.1.11
greenlet==3.1.1
h11==0.14.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
isodate==0.7.2
itsdangerous==2.2.0
jinja2==3.1.6
jmespath==1.0.1
kthread==0.2.3
markdown==3.6
markdown-it-py==3.0.0
markupsafe==3.0.2
marshmallow==3.21.2
mdurl==0.1.2
msal==1.32.0
msal-extensions==1.3.1
networkx==3.3
numpy==2.2.4
openpyxl==3.1.2
ordered-set==4.1.0
packaging==24.2
pandas==2.2.2
passlib==1.7.4
pathlib==1.0.1
playwright==1.51.0
pyarrow==18.1.0
pycparser==2.22 ; platform_python_implementation != 'PyPy'
pyee==12.1.1
pygments==2.19.1
pyjwt==2.10.1
pymongo==4.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.12.0
python-slugify==8.0.4
python-socketio==5.13.0
pytz==2024.1
pyyaml==6.0.2
requests==2.32.3
rich==14.0.0
s3transfer==0.10.4
setuptools==78.1.0
simple-websocket==1.0.0
six==1.17.0
sqlalchemy==2.0.30
taipy==4.0.3
taipy-common==4.0.3
taipy-core==4.0.3
taipy-gui==4.0.3
taipy-rest==4.0.3
taipy-templates==4.0.3
text-unidecode==1.3
toml==0.10.2
twisted==24.7.0
types-python-dateutil==2.9.0.20241206
typing-extensions==4.13.0
tzdata==2025.2
tzlocal==5.2
urllib3==2.4.0
watchdog==4.0.1
werkzeug==3.1.3
wsproto==1.2.0
zope-event==5.0
zope-interface==7.2
